Create the functionality.js in the modules folder and implement the functions to add, edit, and remove
